Rename InstanceStatus class
