+++ src/mysql_shell/clients/instance.py
@@ -8,9 +8,8 @@ from ..builders import StringQueryQuoter
 from ..executors import BaseExecutor
 from ..executors.errors import ExecutionError
 from ..models.account import Role, User
-from ..models.roles import InstanceRole
+from ..models.instance import InstanceRole, InstanceState
 from ..models.statement import LogType, VariableScope
-from ..models.status import InstanceStatus
 
 logger = logging.getLogger()
 
@@ -227,7 +226,7 @@ class MySQLInstanceClient:
         else:
             return [row["cluster_name"] for row in rows]
 
-    def get_instance_replication_state(self) -> InstanceStatus | None:
+    def get_instance_replication_state(self) -> InstanceState | None:
         """Gets the instance replication state."""
         query = (
             "SELECT member_state "
@@ -245,7 +244,7 @@ class MySQLInstanceClient:
             return None
 
         state = rows[0]["member_state"]
-        state = InstanceStatus(state) if state else None
+        state = InstanceState(state) if state else None
         return state
 
     def get_instance_replication_role(self) -> InstanceRole | None:
